pyproject.toml
[project]
name = "betarxiv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.4",
    "anyio>=4.9.0",
    "docling",
    "fastapi[all]",
    "google-genai",
    "numpy",
    "ollama",
    "pgvector",
    "psycopg",
    "pydantic",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
    "pytest>=8.3.5",
    "python-dotenv",
    "twisted>=24.11.0",
    "watchdog",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest-asyncio>=1.0.0",
]
